[#14] refactor: move password and token util functions into seperate files
