+++ backend/user-service/__tests__/routes/user.routes.test.ts
@@ -7,7 +7,7 @@ import mongoose from 'mongoose'
 import { agent, Agent } from 'supertest'
 import logger from '../../src/common/logger.util'
 import connectToDatabase from '../../src/common/mongodb.util'
-import { generateAccessToken } from '../../src/controllers/auth.controller'
+import { generateAccessToken } from '../../src/common/token.util'
 import defaultErrorHandler from '../../src/middlewares/errorHandler.middleware'
 import '../../src/middlewares/passportJwt.middleware'
 import userSchema from '../../src/models/user.model'
